Build packages compatible with multiple numpys
NumPy maintains backwards compatibility and therefore extensions built with a
older version of NumPy will work with newer versions. To take advantage of this
and build a single pandas package that will work with multiple version of
NumPy, the oldest version of NumPy provided for a particular Python
version (Python 3.5 - 1.9 Python 3.6 - 1.11) or in the case of Python 2.7
the minimum version required by Pandas (1.7) is pinned as a build requirement.
The run requirement is then specified as any NumPy version greater than or
equal to this pinned version.