Add numpy as a runtime dependency with the correct lower bound
Follows up on https://github.com/conda-forge/numpy-feedstock/issues/324#issuecomment-2294607709,
which pointed out that the dependency is currently missing and that it isn't
correct to depend only on the `run_exports` of `numpy` itself (that's ABI-only
and allows >=1.19).

The upstream constraints for v2.2.2 are at
https://github.com/pandas-dev/pandas/blob/v2.2.2/pyproject.toml#L30-L32. Note
that (unlike in `pyproject.toml`) it is not necessary to have separate lower
bounds for different Python versions that correspond to the first `numpy`
version with support for that Python version. A single lower bound is enough;
there will not be (for example) `numpy` conda-forge packages for 1.23.0 and
python 3.11. In `pyproject.toml` the constraints have to be more complex only
to avoid `pip` & co trying to build a too-old version from source.